Support fn_capture as lambda.
I think it's safe to make it always a lambda.

Python expressions never cross line boundaries, including in Lambdas.

In general, this means that not all gleam expressions can map to python
expressions (consider case->match, where match is a statement in python,
not an expression). But I *think* it means that any expression that can
map to a python expression can be wrapped in a lambda like this.

The parens make it so the expression is self-contained and you can call
it.